+++ RELEASE_NOTES.md
@@ -1,3 +1,12 @@
+### 5.4.0-prerelease0004
+- extend dropdownConfig on event (allow open menu by hover or click)
+
+### 5.4.0-prerelease0003
+- prevent dropdown to fire init message
+
+### 5.4.0-prerelease0002
+- prevent dropdowns onChange messages (not set be GUI)
+
 ### 5.4.0-prerelease0001
 - Updated Fomantic-UI to 2.8.8
 - Fixed and improved resource management